Make core components more compatible with Storybook
