+++ packages/vulcan-core/lib/modules/components.js
@@ -4,4 +4,9 @@ Also imported by Storybook
 
 */
 
-export { default as Card } from './components/Card.jsx';
+import './components/Avatar.jsx';
+import './components/Card.jsx';
+import './components/Loading.jsx';
+import './components/LoadingButton.jsx';
+import './components/NewButton.jsx';
+import './components/EditButton.jsx';
