+++ packages/vulcan-core/lib/modules/index.js
@@ -10,15 +10,11 @@ export * from './components.js';
 export { default as Layout } from './components/Layout.jsx';
 export { default as App } from './components/App.jsx';
 export { default as Icon } from './components/Icon.jsx';
-export { default as Loading } from './components/Loading.jsx';
 export { default as ShowIf } from './components/ShowIf.jsx';
-export { default as NewButton } from './components/NewButton.jsx';
-export { default as EditButton } from './components/EditButton.jsx';
 export { default as MutationButton } from './components/MutationButton.jsx';
 export { default as Error404 } from './components/Error404.jsx';
 export { default as DynamicLoading } from './components/DynamicLoading.jsx';
 export { default as HeadTags } from './components/HeadTags.jsx';
-export { default as Avatar } from './components/Avatar.jsx';
 export { default as Datatable } from './components/Datatable.jsx';
 export { default as Flash } from './components/Flash.jsx';
 export { default as HelloWorld } from './components/HelloWorld.jsx';
